# Use optimized grouping method for PooledArrays