--- a/REQUIRE
+++ b/REQUIRE
@@ -12,4 +12,4 @@ Compat 0.59.0
 Tables 0.1.14
 IteratorInterfaceExtensions 0.1.1
 TableTraits 0.4.0
-PooledArrays
+PooledArrays 0.5.0