The PR currently adds a dependency on PooledArrays. I'm not sure what's the best approach. We could define a custom method behind `@require PooledArrays` to avoid the dependency (calling a separate function to avoid duplicating the code). But I figured we may want to add a `pool!` function to parallel `categorical!`, in which case we would have to take the dependency anyway. Though we can defer this issue for later, and use `@require` until then.

Most of the work was to adapt tests to make them more generic. I've added a helper for that. Note that it is better to ignore whitespace to see what actually changed.

Obligatory benchmark:
```julia
using DataFrames, BenchmarkTools, PooledArrays

df1 = DataFrame(key1=categorical(string.(rand(1:10, 500_000))),
                key2=categorical(string.(rand(1:10, 500_000))));
df1p = DataFrame(key1=PooledArray(string.(rand(1:10, 500_000))),
                 key2=PooledArray(string.(rand(1:10, 500_000))));

df2 = DataFrame(key1=categorical(string.(rand(1:1000, 500_000))),
                key2=categorical(string.(rand(1:1000, 500_000))));
df2p = DataFrame(key1=PooledArray(string.(rand(1:1000, 500_000))),
                 key2=PooledArray(string.(rand(1:1000, 500_000))));

# master
julia> @btime groupby(df1, :key1);

  2.304 ms (51 allocations: 7.63 MiB)

julia> @btime groupby(df2, :key1);
  4.240 ms (51 allocations: 7.66 MiB)

julia> @btime groupby(df1, [:key1, :key2]);
  3.364 ms (53 allocations: 7.63 MiB)

julia> @btime groupby(df2, [:key1, :key2]);
  37.651 ms (59 allocations: 22.23 MiB)

julia> @btime groupby(df1p, :key1);
  28.769 ms (48 allocations: 15.45 MiB)

julia> @btime groupby(df2p, :key1);
  34.061 ms (48 allocations: 15.46 MiB)

julia> @btime groupby(df1p, [:key1, :key2]);
  58.622 ms (499949 allocations: 23.08 MiB)

julia> @btime groupby(df2p, [:key1, :key2]);
  82.056 ms (106523 allocations: 23.08 MiB)


# This PR
julia> @btime groupby(df1, :key1);
  2.357 ms (51 allocations: 7.63 MiB)

julia> @btime groupby(df2, :key1);
  4.197 ms (51 allocations: 7.66 MiB)

julia> @btime groupby(df1, [:key1, :key2]);
  3.337 ms (53 allocations: 7.63 MiB)

julia> @btime groupby(df2, [:key1, :key2]);
  41.909 ms (59 allocations: 22.23 MiB)


julia> @btime groupby(df1p, :key1);
  2.308 ms (51 allocations: 7.63 MiB)

julia> @btime groupby(df2p, :key1);
  4.163 ms (51 allocations: 7.66 MiB)

julia> @btime groupby(df1p, [:key1, :key2]);
  3.249 ms (53 allocations: 7.63 MiB)

julia> @btime groupby(df2p, [:key1, :key2]);
  37.511 ms (59 allocations: 22.24 MiB)
```