@bkamins 2019-01-29
Assuming PooledArrays.jl will be maintained I think it is OK to have it as a dependency as it is really a small package.
The PR looks good (I have to sleep over it though).