@nalimilan 2019-01-31
Let's wait for https://github.com/JuliaComputing/PooledArrays.jl/pull/15.